+++ src/main/java/com/cotato/kampus/domain/chat/application/ChatRoomDeleter.java
@@ -3,7 +3,7 @@ package com.cotato.kampus.domain.chat.application;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.chat.dao.ChatRoomRepository;
+import com.cotato.kampus.domain.chat.dao.repository.ChatRoomRepository;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
