+++ src/main/java/com/cotato/kampus/domain/chat/application/ChatRoomValidator.java
@@ -3,7 +3,7 @@ package com.cotato.kampus.domain.chat.application;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.chat.dto.ChatRoomDto;
+import com.cotato.kampus.domain.chat.domain.ChatRoomDto;
 import com.cotato.kampus.global.error.ErrorCode;
 import com.cotato.kampus.global.error.exception.AppException;
