+++ src/main/java/com/cotato/kampus/domain/chat/application/ChatroomMetadataDeleter.java
@@ -3,7 +3,7 @@ package com.cotato.kampus.domain.chat.application;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.chat.dao.ChatroomMetadataRepository;
+import com.cotato.kampus.domain.chat.dao.repository.ChatroomMetadataRepository;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
