+++ src/main/java/com/cotato/kampus/domain/chat/application/MessageReadStatusUpdater.java
@@ -3,8 +3,8 @@ package com.cotato.kampus.domain.chat.application;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cotato.kampus.domain.chat.dao.MessageReadStatusRepository;
-import com.cotato.kampus.domain.chat.domain.MessageReadStatus;
+import com.cotato.kampus.domain.chat.dao.repository.MessageReadStatusRepository;
+import com.cotato.kampus.domain.chat.dao.entity.MessageReadStatus;
 
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
