+++ src/main/java/com/cotato/kampus/domain/chat/application/PostChatService.java
@@ -8,16 +8,16 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.cotato.kampus.domain.board.application.BoardFinder;
 import com.cotato.kampus.domain.board.dto.BoardDto;
-import com.cotato.kampus.domain.chat.domain.ChatMessage;
-import com.cotato.kampus.domain.chat.domain.Chatroom;
-import com.cotato.kampus.domain.chat.domain.ChatroomMetadata;
-import com.cotato.kampus.domain.chat.dto.ChatMessageSlice;
-import com.cotato.kampus.domain.chat.dto.ChatMessageSliceSnapshot;
-import com.cotato.kampus.domain.chat.dto.ChatNotification;
-import com.cotato.kampus.domain.chat.dto.ChatNotificationResult;
-import com.cotato.kampus.domain.chat.dto.ChatRoomDetailDto;
-import com.cotato.kampus.domain.chat.dto.ChatRoomPreview;
-import com.cotato.kampus.domain.chat.dto.ChatRoomPreviewList;
+import com.cotato.kampus.domain.chat.dao.entity.ChatMessage;
+import com.cotato.kampus.domain.chat.dao.entity.Chatroom;
+import com.cotato.kampus.domain.chat.dao.entity.ChatroomMetadata;
+import com.cotato.kampus.domain.chat.domain.ChatMessageSlice;
+import com.cotato.kampus.domain.chat.domain.ChatMessageSliceSnapshot;
+import com.cotato.kampus.domain.chat.domain.ChatNotification;
+import com.cotato.kampus.domain.chat.domain.ChatNotificationResult;
+import com.cotato.kampus.domain.chat.domain.ChatRoomDetailDto;
+import com.cotato.kampus.domain.chat.domain.ChatRoomPreview;
+import com.cotato.kampus.domain.chat.domain.ChatRoomPreviewList;
 import com.cotato.kampus.domain.common.application.ApiUserResolver;
 import com.cotato.kampus.domain.post.application.PostFinder;
 import com.cotato.kampus.domain.post.dto.PostDto;
