+++ src/main/java/com/cotato/kampus/domain/chat/dao/repository/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.dao;
+package com.cotato.kampus.domain.chat.dao.repository;
 
 import java.util.Optional;
 
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.cotato.kampus.domain.chat.domain.ChatMessage;
+import com.cotato.kampus.domain.chat.dao.entity.ChatMessage;
 
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
