+++ src/main/java/com/cotato/kampus/domain/chat/dao/repository/ChatRoomRepository.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.dao;
+package com.cotato.kampus.domain.chat.dao.repository;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.cotato.kampus.domain.chat.domain.Chatroom;
+import com.cotato.kampus.domain.chat.dao.entity.Chatroom;
 
 public interface ChatRoomRepository extends JpaRepository<Chatroom, Long> {
