+++ src/main/java/com/cotato/kampus/domain/chat/dao/repository/ChatroomMetadataRepository.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.dao;
+package com.cotato.kampus.domain.chat.dao.repository;
 
 import java.util.Optional;
 
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.cotato.kampus.domain.chat.domain.ChatroomMetadata;
+import com.cotato.kampus.domain.chat.dao.entity.ChatroomMetadata;
 
 public interface ChatroomMetadataRepository extends JpaRepository<ChatroomMetadata, Long> {
