+++ src/main/java/com/cotato/kampus/domain/chat/domain/ChatMessageSlice.java
@@ -1,8 +1,8 @@
-package com.cotato.kampus.domain.chat.dto;
+package com.cotato.kampus.domain.chat.domain;
 
 import org.springframework.data.domain.Slice;
 
-import com.cotato.kampus.domain.chat.domain.ChatMessage;
+import com.cotato.kampus.domain.chat.dao.entity.ChatMessage;
 
 public record ChatMessageSlice(
 	Slice<ChatMessage> chatMessages
