+++ src/main/java/com/cotato/kampus/domain/chat/domain/ChatMessageSliceSnapshot.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.chat.dto;
+package com.cotato.kampus.domain.chat.domain;
 
 import java.util.List;
