+++ src/test/java/com/cotato/kampus/domain/chat/application/ChatRoomAppenderTest.java
@@ -11,8 +11,8 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 
-import com.cotato.kampus.domain.chat.dao.ChatRoomRepository;
-import com.cotato.kampus.domain.chat.domain.Chatroom;
+import com.cotato.kampus.domain.chat.dao.repository.ChatRoomRepository;
+import com.cotato.kampus.domain.chat.dao.entity.Chatroom;
 import com.cotato.kampus.domain.chat.enums.InitiatedFrom;
 
 @ExtendWith(MockitoExtension.class)
